hv: vm: Move shutdown_vm to common scope
The shutdown_vm is renamed to destroy_vm,
and introduce arch_deinit_vm as mandatory arch public API.

The original shutdown_vm checks if all VMs are shutting down,
and shutdown platform if no other VMs are running.

This logic is moved to the caller of original shutdown_vm
except hypercall, as we post-launched VM shutdown should not
trigger platform shutdown.

Tracked-On: #8830
Acked-by: Wang Yu1 <yu1.wang@intel.com>
