+++ hypervisor/common/hypercall.c
@@ -274,7 +274,7 @@ int32_t hcall_destroy_vm(__unused struct acrn_vcpu *vcpu, struct acrn_vm *target
 
 	if (is_paused_vm(target_vm)) {
 		/* TODO: check target_vm guest_flags */
-		ret = shutdown_vm(target_vm);
+		ret = destroy_vm(target_vm);
 	}
 
 	return ret;
